Force Core attr creator to operate on singleton class
Previously, I was executing the attr_<create attr type> on the core
class instead of on the singleton class.

This caused unexpected side effects. It was creating the methods for all
instances of the class (obviously). Thus some methods were already
defined for new instances of the class.

Instead I wanted attrs defined only for a given instance.

I may some day provide core class level attr creation, but for now I
want it to be instance only.