Change the Readme to read more universally
I worry that some people might be upset with my choice of examples. Thus: trees.

I realize some people might be upset that I changed it but I'm not a
brony and I hope that I don't upset anyone.

Also I am not pushing this README to rubygems because I don't think this
merits a release of any kind.